Extract DataListOverflowFade into its own component
