+++ packages/components/src/DataList/components/DataListFilters/components/DataListOverflowFade/DataListOverflowFade.css.d.ts
@@ -1,0 +1,9 @@
+declare const styles: {
+  readonly "fadeContainer": string;
+  readonly "overflowGrid": string;
+  readonly "overflowTrigger": string;
+  readonly "overflowLeft": string;
+  readonly "overflowRight": string;
+};
+export = styles;
+
